handcap_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
nba_api==1.4.1
pandas==2.2.0
numpy==1.26.3